Correction of warnings
git-svn-id: https://svn.salilab.org/imp/trunk@4224 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/em/include/Volume.h
@@ -84,9 +84,9 @@ public:
     std::ifstream in;
     in.open(filename.c_str(), std::ios::in | std::ios::binary);
     header_.read(in,skip_type_check,force_reversed,skip_extra_checkings);
-    data_.resize(header_.get_number_of_slices(),
-                 header_.get_number_of_rows(),
-                 header_.get_number_of_columns());
+    data_.resize((int)header_.get_number_of_slices(),
+                 (int)header_.get_number_of_rows(),
+                 (int)header_.get_number_of_columns());
     data_.read_binary(in);
     in.close();
   }
